move configs and subcomponents to sep files
